Extra filters for item ledger
